added VLRawBankDecoder, to be tested
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@141166 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: fe761f5f0b7ff09551357db93aa7f008acee0360

+++ Lbcom/VL/VLDAQ/src/VLADCWord.h
@@ -4,7 +4,6 @@
 /** @class VLADCWord
  *
  *  Decoding/encoding of VeloLite ADC words
- *  (based on SiADCWord)
  *
  */
 
@@ -25,8 +24,8 @@ public:
   ~VLADCWord() {}
 
   unsigned char value() const {return m_value;}
-  double adc() const {
-    return double((m_value & adcMask) >> adcBits);
+  unsigned int adc() const {
+    return (m_value & adcMask) >> adcBits;
   }
   bool endCluster() const {
     return ((m_value & endMask) >> endBits != 0u) ? true : false;
